+++ src/components/CharacterGenerator/ui/menuScreens/CommonScreen.tsx
