+++ src/components/CharacterGenerator/ui/menuScreens/OriginClassScreen/ClassDescription.tsx
@@ -1,6 +1,6 @@
-import { useCharacter } from "../../../lib/CharacterContext"
-import { useData } from "../../../lib/DataContext";
-import { PassiveDisplay } from "../../common/PassiveDisplay";
+import { useCharacter } from "../../../../../lib/CharacterContext"
+import { useData } from "../../../../../lib/DataContext";
+import { PassiveDisplay } from "../../../../common/PassiveDisplay";
 
 export const ClassDescription = () => {
     const { class: charClass } = useCharacter();
