+++ src/components/CharacterGenerator/ui/menuScreens/OriginClassScreen/OriginClassScreen.tsx
@@ -1,5 +1,5 @@
-import { useCharacter, useCharacterDispatch } from "../../../lib/CharacterContext";
-import { useData } from "../../../lib/DataContext";
+import { useCharacter, useCharacterDispatch } from "../../../../../lib/CharacterContext";
+import { useData } from "../../../../../lib/DataContext";
 import { ClassDescription } from "./ClassDescription";
 
 export const OriginClassScreen = () => {
